+++ canals/component/__init__.py
@@ -1,2 +1,2 @@
-from canals.component.component import name
+from canals.component.component import component
 from canals.component.save_init_params import set_default_component_attributes
