--- test/components/__init__.py
@@ -1,11 +1,0 @@
-from test.components.test_accumulate import Accumulate
-from test.components.test_add_value import AddFixedValue
-from test.components.test_double import Double
-from test.components.test_parity import Parity
-from test.components.test_greet import Greet
-from test.components.test_merge_loop import MergeLoop
-from test.components.test_remainder import Remainder
-from test.components.test_repeat import Repeat
-from test.components.test_subtract import Subtract
-from test.components.test_sum import Sum
-from test.components.test_threshold import Threshold
