+++ test/pipelines/integration/test_linear_pipeline.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from pprint import pprint
 
 from canals.pipeline import Pipeline
-from test.components import AddFixedValue, Double
+from test.test_components import AddFixedValue, Double
 
 import logging
