+++ test/pipelines/integration/test_looping_and_merge_pipeline.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from pprint import pprint
 
 from canals.pipeline import Pipeline
-from test.components import Accumulate, AddFixedValue, Threshold, Sum, MergeLoop
+from test.test_components import Accumulate, AddFixedValue, Threshold, Sum, MergeLoop
 
 import logging
