+++ test/pipelines/integration/test_variable_decision_and_merge_pipeline.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from pprint import pprint
 
 from canals.pipeline import Pipeline
-from test.components import AddFixedValue, Remainder, Double, Sum
+from test.test_components import AddFixedValue, Remainder, Double, Sum
 
 logging.basicConfig(level=logging.DEBUG)
