+++ test/test_components/__init__.py
@@ -1,0 +1,11 @@
+from test.test_components.test_accumulate import Accumulate
+from test.test_components.test_add_value import AddFixedValue
+from test.test_components.test_double import Double
+from test.test_components.test_parity import Parity
+from test.test_components.test_greet import Greet
+from test.test_components.test_merge_loop import MergeLoop
+from test.test_components.test_remainder import Remainder
+from test.test_components.test_repeat import Repeat
+from test.test_components.test_subtract import Subtract
+from test.test_components.test_sum import Sum
+from test.test_components.test_threshold import Threshold
