+++ test/test_components/test_greet.py
@@ -2,13 +2,13 @@ import logging
 
 from dataclasses import dataclass
 
-from canals import name
+from canals import component
 
 
 logger = logging.getLogger(__name__)
 
 
-@name
+@component
 class Greet:
     """
     Logs a greeting message without affecting the value passing on the connection.
